+++ hutool-core/src/test/java/com/xiaoleilu/hutool/core/util/CollectionUtilTest.java
@@ -1,6 +1,7 @@
 package com.xiaoleilu.hutool.core.util;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Set;
 
 import org.junit.Assert;
@@ -27,5 +28,32 @@ public class CollectionUtilTest {
 
 		ArrayList<Object> v2s = CollectionUtil.valuesOfKeys(v2, keys);
 		Assert.assertEquals(v2s, CollectionUtil.newArrayList(15, 13, "李四"));
+	}
+	
+	@Test
+	public void unionTest() {
+		ArrayList<String> list1 = CollectionUtil.newArrayList("a", "b", "b", "c", "d", "x");
+		ArrayList<String> list2 = CollectionUtil.newArrayList("a", "b", "b", "b", "c", "d");
+		
+		Collection<String> union = CollectionUtil.union(list1, list2);
+		Assert.assertEquals("[d, b, b, b, c, a, x]", union.toString());
+	}
+	
+	@Test
+	public void intersectionTest() {
+		ArrayList<String> list1 = CollectionUtil.newArrayList("a", "b", "b", "c", "d", "x");
+		ArrayList<String> list2 = CollectionUtil.newArrayList("a", "b", "b", "b", "c", "d");
+		
+		Collection<String> union = CollectionUtil.intersection(list1, list2);
+		Assert.assertEquals("[d, b, b, c, a]", union.toString());
+	}
+	
+	@Test
+	public void disjunctionTest() {
+		ArrayList<String> list1 = CollectionUtil.newArrayList("a", "b", "b", "c", "d", "x");
+		ArrayList<String> list2 = CollectionUtil.newArrayList("a", "b", "b", "b", "c", "d");
+		
+		Collection<String> disjunction = CollectionUtil.disjunction(list1, list2);
+		Assert.assertEquals("[b, x]", disjunction.toString());
 	}
 }