+++ hutool-poi/src/main/java/com/xiaoleilu/hutool/poi/excel/ExcelUtil.java
@@ -1,0 +1,5 @@
+package com.xiaoleilu.hutool.poi.excel;
+
+public class ExcelUtil {
+
+}